One big code dump, due to most of it being interdependent in some way. And carving out what is not, would mostly affect the time spent on preparing this commit. - Changed the main File-handling routine, all internal commands and all relevant   .gfmmodules to use filesystem locks as opposed to path-strings to identify   directories where possible. - Listers got a new path-gadget that dynamically allocates its string buffer. - Made all path-buffers used by Listers, Directory-buffers and File-handling   functions dynamically allocated, and removed the 255 character path limit on   file-handling. WARNING: This is HIGLY EXPERIMENTAL, there might still be   buffer-overflows lurking in spots I have missed during the rewrite.   Also, Galileo system paths like the program directory, the desktop directory,   paths to button-banks and any paths that are part of the configuration like,   backgrounds, sounds and button images, are still limited to 255 characters. - Changed the Drag'n'Drop routines to stop using am_Reserved of struct   AppMessage. Drag'n'Drop between Galileo-Windows now uses it's own custom   struct where needed. - Incoming AppMessages to Listers or Button banks would be duplicated, then the   WbArgs would be parsed into a list of ArgArrayEntry, which in-turn would be   parsed into a list of ExtarnalEntry, before that was finally parsed into a   list of FunctionEntry to be consumed by the chosen Function. After which all   the lists where un-allocated.   This is now changed to eliminate the ArgArayEntry list. And if the message is   of the new internal type it is just passed on as opposed to copied in the   first step. - AppMessages destined for AppIcons on the Galileo Desktop would be sent to the   AppMessage-port belonging to the desktop, before being forwarded to the port   belonging AppIcon. This is changed to send them directly to the AppIcon port. - Incoming AppMessages to the Desktop, Listers, Button Banks or Program Groups   that has multiple WBArgs with different parent-directories are now handled   correctly. - Replace icon by Drag'n'Drop with Shift+Alt where the destination does not   have an icon now uses GetDefDiskObject() to create one of the right type, as   opposed to blindly copying the source icon. - Changed XADbrowser memory-allocations to use pools. - Moved XADbrovser instance-data from the stack to a separate memory   allocation. - Rewrote XADexteract and the copy command of XADopen to use on common   file-extract function. - For the n'th time changed XADbrowser update of destination lister when   extracting files. - Fixed XADbrowser handling of empty-directories represented by fake files. - Fixed XADbrowser error-requester handling. - Made any Listers showing the DeviceList refresh on disk-change. - Fixed the bug I introduced in commit 7419662, that made Listers showing   the DeviceList switch to showing the Galileo program-directory if locale   settings was changed. - When viewing the DeviceList in icon mode. It would first render all the icons   without position-info, on top of each-other in the top left corner of the   window. Then it would clear the window and render them again laid out   correctly. And in some cases it would stop after the first render.   This is no changed to one single rendering pass. - Fixed DeviceList handling of multiple disks with the same name in   icon-modes. - Fixed DeviceList sometimes shrinking the window to minimum size when   switching from text-mode to icon-modes. - Clearing the the DirList area of a text-mode lister before showing a new   directory was done by rendering an empty DirList, line by line. This is now   changed to use EraseRect(). - Made desktop left-outs disk-change aware. - Made desktop left-outs and group items use the volume DateStamp to   identify the originating disk. - Made the position records for lister windows use the volume DateStamp   to identify the disk. - When disks with identical names are present, and directory is changed with   the lister path-gadget, it will stay on the current disk as long as the   volume name was not changed. - Moved the main library LibData to the near-data section. - Stored a copy of the main library LibBase-pointer in far-data, and added a   simple assembler trampoline to set a6 from that in front of all callbacks in   the main library. To let them use the global library bases normally, and get   rid of most cases of "#define SomeLibBase SomeStruct->StoredLibBase". - The LiBase changes left a flood of "Warning: 88" because I changed some of the   functions to take "struct Library *GalileoFMBase" in register a6 were they   previously took "struct MyLibrary *libbase" or the LibBase was implicit.   To get rid of those I have to convert all the other "struct MyLibrary *libbase"   the same way, and probably also most/all of the functions where the a6 is   implicit. To limit the scope of this commit and finally get it out the door,   warning 88 is disabled for the main library for now. - According to the SAS/C manual many of the function pointer prototypes were   wrong as the "__asm" keyword was in front of the "(*)". This is now changed to   be in accordance with the documentation. - Removed L_SaveSettings() from the main library as it was an unused leftover   from earlier versions. - Disabled compiling of "Program/backdrop_utils.c:backdrop_find_disk()" and   "Program/desktop.c:desktop_snapshot_icon()" as they are currently not used   anywhere. - Reorganized the includes in "Program/" trying to make each file depends only   on the includes it actually needs. - Integrated stack-usage checking from   https://aminet.net/package/dev/misc/StackCheck by Günther Röhrich for the   debug-build. - ARexx, callback and library function additions/changes, see the changed files   in "Documentation" and "SDK/docs". - Various changes needed for the local debug-build. - Possibly some stuff I forgot about.. - And almost certainly some new bugs.

+++ Source/Modules/update.c
@@ -37,9 +37,10 @@ For more information on Directory Opus for Windows please see:
 */
 
 #include "update.h"
-#include "/Program/position.h"
 
 char *version="$VER: update.gfmmodule 0.2 "__AMIGADATE__" ";
+
+BOOL alloc_fail = FALSE;
 
 int __asm __saveds L_Module_Entry(
 	register __a0 char *args,
@@ -143,8 +144,18 @@ int __asm __saveds L_Module_Entry(
 		flags|=UPDATEF_DONE_POSITION;
 	}
 
+	// Haven't done iffchunk?
+	if (!(flags&UPDATEF_DONE_IFFCHUNK))
+	{
+		// Open status window
+		if (!window) window=open_status(screen);
+
+		if (update_iffchunk())
+		    flags|=UPDATEF_DONE_IFFCHUNK;
+	}
+
 	// Haven't done leftouts?
-	if (!(flags&UPDATEF_DONE_LEFTOUTS))
+	if (!alloc_fail && !(flags&UPDATEF_DONE_LEFTOUTS))
 	{
 		struct List list;
 		APTR memory;
@@ -165,7 +176,7 @@ int __asm __saveds L_Module_Entry(
 	}
 
 	// Haven't done groups?
-	if (!(flags&UPDATEF_DONE_GROUPS) || !(flags&UPDATEF_DONE_FIXGROUPS))
+	if (!alloc_fail && !(flags&UPDATEF_DONE_GROUPS) || !(flags&UPDATEF_DONE_FIXGROUPS))
 	{
 		// Open status window
 		if (!window) window=open_status(screen);
@@ -175,7 +186,7 @@ int __asm __saveds L_Module_Entry(
 	}
 
 	// Haven't done storage?
-	if (!(flags&UPDATEF_DONE_STORAGE))
+	if (!alloc_fail && !(flags&UPDATEF_DONE_STORAGE))
 	{
 		// Open status window
 		if (!window) window=open_status(screen);
@@ -185,7 +196,7 @@ int __asm __saveds L_Module_Entry(
 	}
 
 	// Haven't done filetypes?
-	if (!(flags&UPDATEF_DONE_FILETYPES))
+	if (!alloc_fail && !(flags&UPDATEF_DONE_FILETYPES))
 	{
 		// Open status window
 		if (!window) window=open_status(screen);
@@ -200,18 +211,8 @@ int __asm __saveds L_Module_Entry(
 		flags|=UPDATEF_DONE_FILETYPES;
 	}
 
-	// Haven't done iffchunk?
-	if (!(flags&UPDATEF_DONE_IFFCHUNK))
-	{
-		// Open status window
-		if (!window) window=open_status(screen);
-
-		if (update_iffchunk())
-		    flags|=UPDATEF_DONE_IFFCHUNK;
-	}
-
 	// Haven't done path keys
-	if (!(flags&UPDATEF_DONE_PATH_KEYS))
+	if (!alloc_fail && !(flags&UPDATEF_DONE_PATH_KEYS))
 	{
 		struct List list;
 		APTR memory;
@@ -236,7 +237,7 @@ int __asm __saveds L_Module_Entry(
 		flags|=UPDATEF_DONE_PATH_KEYS;
 	}
 	else
-	if (!(flags&UPDATEF_DONE_PATHFORMAT))
+	if (!alloc_fail && !(flags&UPDATEF_DONE_PATHFORMAT))
 	{
 	    // Delete pathformat module
 	    DeleteFile("PROGDIR:modules/pathformat.gfmmodule");
@@ -247,7 +248,7 @@ int __asm __saveds L_Module_Entry(
 	}
 
 	// Haven't done theme stuff
-	if (!(flags&UPDATEF_DONE_THEMES))
+	if (!alloc_fail && !(flags&UPDATEF_DONE_THEMES))
 	{
 		// Create sounds directory
 		if (file=CreateDir("PROGDIR:Sounds"))
@@ -267,8 +268,23 @@ int __asm __saveds L_Module_Entry(
 		flags|=UPDATEF_DONE_THEMES;
 	}
 
+	// Haven't done disk-id positions?
+	if (!alloc_fail && !(flags&UPDATEF_DONE_POSITOPOSR))
+	{
+		struct List list;
+
+		// Open status window
+		if (!window) window=open_status(screen);
+
+		// Read position-info file
+		if (update_position_info(&list))
+		{
+		    flags|=UPDATEF_DONE_POSITOPOSR;
+		}
+	}
+
 	// Have flags changed?
-	if (oldflags!=flags)
+	if (!alloc_fail && oldflags!=flags)
 	{
 		// Set bits to 'rwed'
 		SetProtection("PROGDIR:system/update-history",0);
@@ -303,12 +319,91 @@ struct Window *open_status(struct Screen *screen)
 }
 
 
+BOOL write_position_info(struct List *list)
+{
+    APTR iff;
+    struct Node *node;
+    BOOL succ = TRUE;
+
+    // Open file
+    if (!(iff=IFFOpen("PROGDIR:System/position-info",IFF_WRITE,ID_GILO)))
+	    return 0;
+
+    // Go through entries
+    for (node=list->lh_Head;node->ln_Succ;node=node->ln_Succ)
+    {
+	    // Valid path, or AppIcon?
+	    if ((node->ln_Name && *node->ln_Name) ||
+		    node->ln_Type==PTYPE_APPICON)
+	    {
+		    ULONG id=0;
+		    short size=0;
+
+		    // Position?
+		    if (node->ln_Type==PTYPE_POSITION)
+		    {
+		        if (node->ln_Pri == 1)
+		        {
+			    // Get id and chunk size
+			    id=ID_POSR;
+			    size=sizeof(position_rec)-sizeof(struct Node);
+		        }
+		        else
+		        {
+			    // Get id and chunk size
+			    id=ID_POSI;
+			    size=sizeof(position_rec_old)-sizeof(struct Node);
+		        }
+		    }
+
+		    // Left-out?
+		    else
+		    if (node->ln_Type==PTYPE_LEFTOUT)
+		    {
+		        if (node->ln_Pri == 1)
+		        {
+			    // Get id and chunk size
+			    id=ID_LFTO;
+			    size=sizeof(leftout_record)-sizeof(struct Node);
+		        }
+		        else
+		        {
+			    // Get id and chunk size
+			    id=ID_LOUT;
+			    size=sizeof(leftout_record_old)-sizeof(struct Node);
+		        }
+		    }
+
+		    // AppIcon?
+		    else
+		    if (node->ln_Type==PTYPE_APPICON)
+		    {
+			    // Get id and chunk size
+			    id=ID_ICON;
+			    size=sizeof(appicon_record)-sizeof(struct Node);
+		    }
+
+		    // Write chunk
+		    if (!(IFFWriteChunk(iff,(APTR)(node+1),id,size+((node->ln_Name)?strlen(node->ln_Name):0))))
+		    {
+			succ = FALSE;
+			break;
+		    }
+	    }
+    }
+
+    // Close file
+    IFFClose(iff);
+
+    return succ;
+}
+
 // Convert leftouts
 BOOL update_do_leftouts(struct List *list,APTR memory)
 {
 	struct DosList *dos;
 	short count=0;
-	BOOL asked=0,dowb=1;
+	BOOL asked=0,dowb=1, succ = FALSE;
 
 	// Lock dos list
 	dos=LockDosList(LDF_VOLUMES|LDF_READ);
@@ -385,60 +480,11 @@ BOOL update_do_leftouts(struct List *list,APTR memory)
 	// Added anything?
 	if (count)
 	{
-		APTR iff;
-		struct Node *node;
-
-		// Open file
-		if (!(iff=IFFOpen("PROGDIR:System/position-info",IFF_WRITE,ID_GILO)))
-			return 0;
-
-		// Go through entries
-		for (node=list->lh_Head;node->ln_Succ;node=node->ln_Succ)
-		{
-			// Valid path, or AppIcon?
-			if ((node->ln_Name && *node->ln_Name) ||
-				node->ln_Type==PTYPE_APPICON)
-			{
-				ULONG id=0;
-				short size=0;
-
-				// Position?
-				if (node->ln_Type==PTYPE_POSITION)
-				{
-					// Get id and chunk size
-					id=ID_POSI;
-					size=sizeof(position_rec)-sizeof(struct Node);
-				}
-
-				// Left-out?
-				else
-				if (node->ln_Type==PTYPE_LEFTOUT)
-				{
-					// Get id and chunk size
-					id=ID_LOUT;
-					size=sizeof(leftout_record)-sizeof(struct Node);
-				}
-
-				// AppIcon?
-				else
-				if (node->ln_Type==PTYPE_APPICON)
-				{
-					// Get id and chunk size
-					id=ID_ICON;
-					size=sizeof(leftout_record)-sizeof(struct Node);
-				}
-
-				// Write chunk
-				if (!(IFFWriteChunk(iff,(APTR)(node+1),id,size+((node->ln_Name)?strlen(node->ln_Name):0))))
-					break;
-			}
-		}
-
-		// Close file
-		IFFClose(iff);
+	    if (write_position_info(list))
+		succ = TRUE;
 	}
 
-	return 1;
+	return succ;
 }
 
 
@@ -455,7 +501,7 @@ BOOL update_iffchunk(void)
 	anchor->ap_Strlen=256;
 
 	// Start match
-	error=MatchFirst("PROGDIR:(Buttons|Commands|Environment|Settings|Filetypes|Storage)/#?",anchor);
+	error=MatchFirst("PROGDIR:(Buttons|Commands|Environment|Settings|Filetypes|Storage|System)/#?",anchor);
 	while (!error)
 	{
 		// Directory?
@@ -918,6 +964,138 @@ void update_envarc_do(char *path)
 }
 
 
+// Upgrade position-info file
+BOOL update_position_info(struct List *list)
+{
+    APTR memory;
+    APTR iff;
+    BOOL succ = FALSE;
+
+    // Allocate memory handle
+    if (!(memory=NewMemHandle(4096,1024,MEMF_CLEAR)))
+    {
+	alloc_fail = TRUE;
+	return 0;
+    }
+
+    // Initialise list
+    NewList(list);
+
+    // Open position info file
+    if (iff=IFFOpen("PROGDIR:System/position-info",IFF_READ,ID_GILO))
+    {
+	ULONG id;
+
+	// Read file
+	while (id=IFFNextChunk(iff,0))
+	{
+	    struct Node *entry;
+	    ULONG size;
+
+	    // Valid chunk?
+	    if (id!=ID_POSI && id!=ID_POSR && id!=ID_LOUT && id!=ID_LFTO && id!=ID_ICON && id!=ID_STRT)
+		    continue;
+
+	    // Allocate record
+	    if (!(entry=AllocMemH(memory, size = IFFChunkSize(iff) + sizeof(struct Node))))
+		goto posupg_fail;
+
+	    // Read chunk data
+	    IFFReadChunkBytes(iff,(APTR)(entry+1),-1);
+
+	    // Set type and name pointer
+	    if (id==ID_POSR)
+	    {
+		entry->ln_Type=PTYPE_POSITION;
+		entry->ln_Name=((position_rec *)entry)->name;
+	    }
+	    else
+	    if (id==ID_POSI)
+	    {
+		//UBYTE pos = 0;
+
+		position_rec *upgrade_entry;
+		position_rec_old *org_entry = (position_rec_old *)entry;
+
+		size += sizeof(position_rec) - sizeof(position_rec_old);
+
+		if (!(upgrade_entry = AllocMemH(memory, size)))
+		    goto posupg_fail;
+
+		upgrade_entry->node.ln_Type = PTYPE_POSITION;
+		upgrade_entry->node.ln_Pri  = 1;
+		upgrade_entry->node.ln_Name = upgrade_entry->name;
+		upgrade_entry->type         = PPTYPE_NEEDDATE;
+		upgrade_entry->format       = org_entry->format;
+		upgrade_entry->icon_x       = org_entry->icon_x;
+		upgrade_entry->icon_y       = org_entry->icon_y;
+		upgrade_entry->text_dims    = org_entry->text_dims;
+		upgrade_entry->icon_dims    = org_entry->icon_dims;
+		upgrade_entry->code         = org_entry->code;
+		upgrade_entry->qual         = org_entry->qual;
+		upgrade_entry->qual_mask    = org_entry->qual_mask;
+		upgrade_entry->qual_same    = org_entry->qual_same;
+		upgrade_entry->flags        = org_entry->flags;
+		strcpy(upgrade_entry->name, org_entry->name);
+#if 0
+		do
+		{
+		    pos++;
+		} while (pos < 32 && upgrade_entry->name[pos] && upgrade_entry->name[pos] != ':');
+
+		if (upgrade_entry->name[pos] != ':')
+		    upgrade_entry->vol_name_len = pos;
+#endif
+		FreeMemH(entry);
+		entry = (struct Node *)upgrade_entry;
+		entry->ln_Name=((position_rec *)entry)->name;
+	    }
+	    else
+	    if (id==ID_LFTO)
+	    {
+		entry->ln_Type=PTYPE_LEFTOUT;
+		entry->ln_Pri=1;
+		entry->ln_Name=((leftout_record *)entry)->name;
+	    }
+	    else
+	    if (id==ID_LOUT)
+	    {
+		entry->ln_Type=PTYPE_LEFTOUT;
+		entry->ln_Name=((leftout_record_old *)entry)->name;
+	    }
+	    else
+	    if (id==ID_ICON)
+	    {
+		entry->ln_Type=PTYPE_APPICON;
+		entry->ln_Name=((leftout_record *)entry)->name;
+	    }
+	    // Add to list
+	    AddTail(list,entry);
+	}
+
+	// Close file
+	IFFClose(iff);
+    }
+
+    if (!(IsListEmpty(list)))
+    {
+	if (write_position_info(list))
+	    succ = TRUE;
+    }
+
+    FreeMemHandle(memory);
+
+    return succ;
+
+posupg_fail:
+    FreeMemHandle(memory);
+
+    alloc_fail = TRUE;
+
+    return 0;
+}
+
+
 // Read position-info file
 APTR read_position_info(struct List *list)
 {
@@ -942,7 +1120,7 @@ APTR read_position_info(struct List *list)
 			struct Node *entry;
 
 			// Valid chunk?
-			if (id!=ID_POSI && id!=ID_LOUT && id!=ID_ICON && id!=ID_STRT)
+			if (id!=ID_POSI && id!=ID_POSR && id!=ID_LOUT && id!=ID_LFTO && id!=ID_ICON && id!=ID_STRT)
 				continue;
 
 			// Allocate record
@@ -953,16 +1131,30 @@ APTR read_position_info(struct List *list)
 			IFFReadChunkBytes(iff,(APTR)(entry+1),-1);
 
 			// Set type and name pointer
+			if (id==ID_POSR)
+			{
+				entry->ln_Type=PTYPE_POSITION;
+				entry->ln_Pri=1;
+				entry->ln_Name=((position_rec *)entry)->name;
+			}
+			else
 			if (id==ID_POSI)
 			{
 				entry->ln_Type=PTYPE_POSITION;
-				entry->ln_Name=((position_rec *)entry)->name;
+				entry->ln_Name=((position_rec_old *)entry)->name;
+			}
+			else
+			if (id==ID_LFTO)
+			{
+				entry->ln_Type=PTYPE_LEFTOUT;
+				entry->ln_Pri=1;
+				entry->ln_Name=((leftout_record *)entry)->name;
 			}
 			else
 			if (id==ID_LOUT)
 			{
 				entry->ln_Type=PTYPE_LEFTOUT;
-				entry->ln_Name=((leftout_record *)entry)->name;
+				entry->ln_Name=((leftout_record_old *)entry)->name;
 			}
 			else
 			if (id==ID_ICON)
@@ -993,7 +1185,7 @@ void update_do_pathkeys(struct List *list)
 	if (IsListEmpty(list)) return;
 
 	// Try to open hotkeys bank
-	if (!(bank=OpenButtonBank("PROGDIR:buttons/hotkeys")))
+	if (!(bank=OpenButtonBank("PROGDIR:buttons/hotkeys", NULL)))
 	{
 		// Create a new bank
 		if (!(bank=NewButtonBank(0,0))) return;
